Restart after a failure (#5)
The default value for RestartSec is 100 ms. This seems way too short in
the case of FlexNet. Experience shows the lmstart script provided with
MATLAB may fail to restart FlexNet if called too soon after FlexNet is
stopped, as ports may still be in use. The delay may be larger than 60s.
To avoid reaching the rate limit StartLimitBurst, we need the value of
RestartSec to be similar to that delay.

See discussion here:
https://unix.stackexchange.com/questions/289629/systemd-restart-always-is-not-honored#324297

A value of 30s doesn't sound that bad. Checked-out licenses will remain
active. Only checking out new licenses will be deferred.

Co-authored-by: Jeff Stafford <jeff.stafford@protonmail.com>